refactor: change file about modal
